Rearrange testUtil folder
