+++ src/components/Button/Button.test.tsx
@@ -1,5 +1,5 @@
 import { screen } from "@testing-library/react";
-import { renderWithProviders } from "../../testUtils";
+import { renderWithProviders } from "../../testUtil/renderWithProviders";
 import Button from "./Button";
 
 describe("Given a Button component", () => {
