+++ src/hooks/useUser.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from "@testing-library/react";
 import jwtDecode from "jwt-decode";
 import { Wrapper } from "../mocks/Wrapper";
-import { store } from "../store";
+import { store } from "../store/store";
 import { User } from "../store/features/users/userSlice/types";
 import { loginUserActionCreator } from "../store/features/users/userSlice/userSlice";
 import { CustomTokenPayload, UserCredentials } from "./types";
