--- src/store/index.ts
@@ -1,17 +1,0 @@
-import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit";
-import { userReducer } from "./features/users/userSlice/userSlice";
-
-export const store = configureStore({
-  reducer: {
-    user: userReducer,
-  },
-});
-
-export type AppDispatch = typeof store.dispatch;
-export type RootState = ReturnType<typeof store.getState>;
-export type AppThunk<ReturnType = void> = ThunkAction<
-  ReturnType,
-  RootState,
-  unknown,
-  Action<string>
->;
