+++ src/store/store.ts
@@ -1,0 +1,25 @@
+import {
+  Action,
+  combineReducers,
+  configureStore,
+  PreloadedState,
+  ThunkAction,
+} from "@reduxjs/toolkit";
+import { userReducer } from "./features/users/userSlice/userSlice";
+
+const rootReducer = combineReducers({ user: userReducer });
+
+export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
+  return configureStore({ reducer: rootReducer, preloadedState });
+};
+
+export const store = setupStore();
+
+export type AppDispatch = typeof store.dispatch;
+export type RootState = ReturnType<typeof rootReducer>;
+export type AppThunk<ReturnType = void> = ThunkAction<
+  ReturnType,
+  RootState,
+  unknown,
+  Action<string>
+>;
