+++ src/testUtil/renderWithProviders.tsx
@@ -1,22 +1,11 @@
-import {
-  combineReducers,
-  configureStore,
-  PreloadedState,
-} from "@reduxjs/toolkit";
+import { PreloadedState } from "@reduxjs/toolkit";
 import { render } from "@testing-library/react";
 import { PropsWithChildren } from "react";
 import { Provider } from "react-redux";
 import { ThemeProvider } from "styled-components";
-import { RootState, store } from "./store";
-import { userReducer } from "./store/features/users/userSlice/userSlice";
-import GlobalStyles from "./styles/GlobalStyles";
-import { theme } from "./styles/theme";
-
-const rootReducer = combineReducers({ user: userReducer });
-
-const setupStore = (preloadedState?: PreloadedState<RootState>) => {
-  return configureStore({ reducer: rootReducer, preloadedState });
-};
+import { RootState, setupStore, store } from "../store/store";
+import GlobalStyles from "../styles/GlobalStyles";
+import { theme } from "../styles/theme";
 
 export const renderWithProviders = (
   ui: React.ReactElement,
